requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents
boto3
requests
mcp